feat: modularize command handling by introducing dedicated modules for billing, markets, admin, and utility commands
